fix: Split geo json component
