+++ src/landscape/components/LandscapeBoundaries.js
@@ -1,29 +1,14 @@
-import React, { useCallback, useState, useEffect } from 'react';
+import React, { useCallback, useState } from 'react';
 import _ from 'lodash/fp';
-import { useSelector, useDispatch } from 'react-redux';
-import { useParams, useNavigate } from 'react-router-dom';
 import { useDropzone } from 'react-dropzone';
-import {
-  Alert,
-  Button,
-  Grid,
-  Link,
-  Paper,
-  Stack,
-  Typography,
-} from '@mui/material';
+import { Alert, Link, Paper, Stack, Typography } from '@mui/material';
 import { Trans, useTranslation } from 'react-i18next';
 
 import theme from 'theme';
 import { GEOJSON_MAX_SIZE } from 'config';
-import { fetchLandscapeForm, saveLandscape } from 'landscape/landscapeSlice';
 import { isValidGeoJson } from 'landscape/landscapeUtils';
-import { useDocumentTitle } from 'common/document';
 import { sendToRollbar } from 'monitoring/logger';
-import PageContainer from 'layout/PageContainer';
 import LandscapeMap from './LandscapeMap';
-import PageHeader from 'layout/PageHeader';
-import PageLoader from 'layout/PageLoader';
 import InlineHelp from 'common/components/InlineHelp';
 
 const openFile = file =>
@@ -172,62 +157,14 @@ const DropZone = props => {
   );
 };
 
-const LandscapeBoundaries = () => {
-  const dispatch = useDispatch();
+const LandscapeBoundaries = props => {
   const { t } = useTranslation();
-  const navigate = useNavigate();
-  const { slug } = useParams();
-  const { fetching, landscape, success } = useSelector(
-    state => state.landscape.form
-  );
-  const [areaPolygon, setAreaPolygon] = useState();
-  const onFileSelected = areaPolygon => {
-    setAreaPolygon(areaPolygon);
-  };
-
-  useDocumentTitle(
-    t('landscape.boundaries_document_title', {
-      name: _.get('name', landscape),
-    }),
-    fetching
-  );
-
-  useEffect(() => {
-    dispatch(fetchLandscapeForm(slug));
-  }, [dispatch, slug]);
-
-  useEffect(() => {
-    if (success) {
-      navigate(`/landscapes/${slug}`);
-    }
-  }, [success, slug, navigate, dispatch]);
-
-  if (fetching) {
-    return <PageLoader />;
-  }
-
-  const onSave = () => {
-    dispatch(
-      saveLandscape({
-        id: landscape.id,
-        areaPolygon,
-      })
-    );
-  };
+  const { areaPolygon, onFileSelected } = props;
 
   return (
-    <PageContainer>
-      <PageHeader
-        header={t('landscape.boundaries_title', {
-          name: _.get('name', landscape),
-        })}
-      />
+    <>
       <Paper variant="outlined" sx={{ padding: 2 }}>
-        <LandscapeMap
-          landscape={{
-            areaPolygon: areaPolygon || _.get('areaPolygon', landscape),
-          }}
-        />
+        <LandscapeMap landscape={{ areaPolygon }} />
         <DropZone onFileSelected={onFileSelected} />
       </Paper>
       <InlineHelp
@@ -249,25 +186,7 @@ const LandscapeBoundaries = () => {
           },
         ]}
       />
-      <Grid
-        container
-        direction="row"
-        justifyContent="space-between"
-        sx={{ marginTop: 2 }}
-      >
-        <Button
-          variant="contained"
-          disabled={!areaPolygon}
-          sx={{ paddingLeft: 5, paddingRight: 5 }}
-          onClick={onSave}
-        >
-          {t('landscape.boundaries_save')}
-        </Button>
-        <Button variant="text" onClick={() => navigate(`/landscapes/${slug}`)}>
-          {t('landscape.boundaries_cancel')}
-        </Button>
-      </Grid>
-    </PageContainer>
+    </>
   );
 };
